pyproject.toml
[tool.poetry]
name = "i2c-1602-display"
version = "0.1.0"
description = ""
authors = ["ntnkiev <ntn.kiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hid = "^1.0.6"
pyusb = "^1.2.1"
hidapi = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
